+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   * `setTool(IIngredient, int)` - use `addTool(IIngredient, int)` instead
 * Changed: updated de_de.lang (PR#58 Xaikii)
 * Fixed: JEI recipe transfer issue with some converted oreDict ingredients
+* Fixed: stacks in secondary output slots are not dropped when the block is broken
 * Note: you will need to refresh the config for the worktable module
 
 1.12.21